# openapi.yaml
openapi: 3.0.3
info:
  title: LIS Labo-Anapath API
  description: API REST de gestion de laboratoire d'anatomopathologie
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Développement local
  - url: https://api.labo-anapath.com/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /roles:
    get:
      tags: [Roles]
      summary: Liste des rôles
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePageApiResponse'
    post:
      tags: [Roles]
      summary: Créer un rôle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleApiResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /roles/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [Roles]
      summary: Détail d'un rôle
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Roles]
      summary: Modifier un rôle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleApiResponse'
    delete:
      tags: [Roles]
      summary: Supprimer un rôle
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
  /roles/{id}/permissions:
    parameters:
      - $ref: '#/components/parameters/id'
    post:
      tags: [Roles]
      summary: Assigner des permissions à un rôle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignPermissionsRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleApiResponse'

  # ── PATIENTS ─────────────────────────────────────────────────────────────────
...
  /reports/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [Reports]
      summary: Détail d'un compte rendu
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Reports]
      summary: Modifier un compte rendu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportApiResponse'
    delete:
      tags: [Reports]
      summary: Supprimer un compte rendu
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
...
  /title-reports:
    get:
      tags: [Title Reports]
      summary: Liste des titres de comptes rendus
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitleReportPageApiResponse'
    post:
      tags: [Title Reports]
      summary: Créer un titre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimpleNameRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitleReportApiResponse'
  /title-reports/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [Title Reports]
      summary: Détail
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitleReportApiResponse'
    put:
      tags: [Title Reports]
      summary: Modifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimpleNameRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitleReportApiResponse'
    delete:
      tags: [Title Reports]
      summary: Supprimer
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'

  # ── PATHOLOGY MACROS ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── REUSABLE PARAMETERS ──────────────────────────────────────────────────────
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
    size:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  responses:
    EmptySuccess:
      description: Opération réussie
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EmptyApiResponse'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
...
    Conflict:
      description: Ressource déjà existante
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
...
  schemas:
...
    ReportStatus:
      type: string
      enum: [DRAFT, PENDING_REVIEW, VALIDATED, DELIVERED]
...
    PageMeta:
      type: object
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 100
        totalPages:
          type: integer
          example: 5
        last:
          type: boolean
    EmptyApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OK
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
    ErrorApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Ressource introuvable
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    SimpleNameRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Hématologie
    AssignPermissionsRequest:
      type: object
      required: [permissionIds]
      properties:
        permissionIds:
          type: array
          items:
            type: string
            format: uuid

    # -- Auth -----------------------------------------------------------------
...
    PermissionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
    RoleRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        permissionIds:
          type: array
          items:
            type: string
            format: uuid
    RoleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionResponse'
        createdAt:
          type: string
          format: date-time
    RoleApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          $ref: '#/components/schemas/RoleResponse'
        timestamp: {type: string, format: date-time}
    RolePageApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                content:
                  type: array
                  items:
                    $ref: '#/components/schemas/RoleResponse'
        timestamp: {type: string, format: date-time}

    # -- Patient --------------------------------------------------------------
...
    ReportRequest:
      type: object
      required: [testOrderId]
      properties:
        testOrderId:
          type: string
          format: uuid
        titleId:
          type: string
          format: uuid
        content:
          type: string
        comment:
          type: string
        reviewedById:
          type: string
          format: uuid
        signatory1Id:
          type: string
          format: uuid
        signatory2Id:
          type: string
          format: uuid
        signatory3Id:
          type: string
          format: uuid
        receiverName:
          type: string
        tagIds:
          type: array
          items:
            type: string
            format: uuid
    ReportResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        testOrderId:
          type: string
          format: uuid
        testOrderCode:
          type: string
        titleId:
          type: string
          format: uuid
        titleName:
          type: string
        content:
          type: string
        comment:
          type: string
        status:
          $ref: '#/components/schemas/ReportStatus'
        isDelivered:
          type: boolean
        receiverName:
          type: string
        branchId:
          type: string
          format: uuid
        tagNames:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReportApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          $ref: '#/components/schemas/ReportResponse'
        timestamp: {type: string, format: date-time}
...
    TitleReportResponse:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        branchId: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
    TitleReportApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          $ref: '#/components/schemas/TitleReportResponse'
        timestamp: {type: string, format: date-time}
    TitleReportPageApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                content:
                  type: array
                  items:
                    $ref: '#/components/schemas/TitleReportResponse'
        timestamp: {type: string, format: date-time}
...